move stream retry testing
